handle queue pending state and livekit ut
