+++ packages/sdk/tests/unit.test.ts
@@ -325,166 +325,6 @@ describe("Decart SDK", () => {
       });
     });
   });
-
-  describe("connect startup conditioning", () => {
-    it("sends initial image with the initial prompt after joining the LiveKit room", async () => {
-      const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-      const sentMessages: Array<{ type: string }> = [];
-
-      class FakeWebSocket {
-        static OPEN = 1;
-        static CLOSED = 3;
-        readyState = FakeWebSocket.OPEN;
-        onopen: (() => void) | null = null;
-        onmessage: ((event: { data: string }) => void) | null = null;
-        onerror: (() => void) | null = null;
-        onclose: ((event: { code: number; reason: string }) => void) | null = null;
-
-        constructor(_url: string) {
-          setTimeout(() => this.onopen?.(), 0);
-        }
-
-        send(data: string): void {
-          const msg = JSON.parse(data) as { type: string };
-          sentMessages.push(msg);
-          if (msg.type === "livekit_join") {
-            setTimeout(
-              () =>
-                this.onmessage?.({
-                  data: JSON.stringify({
-                    type: "livekit_room_info",
-                    livekit_url: "wss://livekit.example.com",
-                    token: "token",
-                    room_name: "room",
-                  }),
-                }),
-              0,
-            );
-          }
-        }
-
-        close(): void {
-          this.readyState = FakeWebSocket.CLOSED;
-          this.onclose?.({ code: 1000, reason: "" });
-        }
-      }
-
-      vi.stubGlobal("WebSocket", FakeWebSocket as unknown as typeof WebSocket);
-
-      try {
-        const connection = new LiveKitConnection({
-          initialImage: "base64-image",
-          initialPrompt: { text: "Anime", enhance: false },
-        });
-        const internal = connection as unknown as {
-          joinRoom: (info: unknown) => Promise<void>;
-        };
-        const joinSpy = vi.spyOn(internal, "joinRoom").mockResolvedValue(undefined);
-        const imageSpy = vi.spyOn(connection, "setImageBase64").mockResolvedValue(undefined);
-
-        await connection.connect("wss://example.com/v1/stream", {} as MediaStream, 750);
-
-        expect(sentMessages).toContainEqual({ type: "livekit_join" });
-        expect(joinSpy).toHaveBeenCalledWith({
-          type: "livekit_room_info",
-          livekit_url: "wss://livekit.example.com",
-          token: "token",
-          room_name: "room",
-        });
-        expect(imageSpy).toHaveBeenCalledWith("base64-image", { prompt: "Anime", enhance: false });
-        expect(connection.state).toBe("connected");
-
-        joinSpy.mockRestore();
-        imageSpy.mockRestore();
-      } finally {
-        vi.unstubAllGlobals();
-      }
-    });
-
-    it("surfaces server error control messages", async () => {
-      const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-      const onError = vi.fn();
-      const connection = new LiveKitConnection({ onError });
-      const internal = connection as unknown as {
-        handleControlMessage: (msg: { type: "error"; error: string }) => void;
-      };
-
-      internal.handleControlMessage({ type: "error", error: "server unavailable" });
-
-      expect(onError).toHaveBeenCalledTimes(1);
-      expect(onError.mock.calls[0][0]).toMatchObject({
-        message: "server unavailable",
-        source: "server",
-      });
-    });
-
-    it("aborts connect when a server error arrives during startup conditioning", async () => {
-      const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-
-      class FakeWebSocket {
-        static OPEN = 1;
-        static CLOSED = 3;
-        readyState = FakeWebSocket.OPEN;
-        onopen: (() => void) | null = null;
-        onmessage: ((event: { data: string }) => void) | null = null;
-        onerror: (() => void) | null = null;
-        onclose: ((event: { code: number; reason: string }) => void) | null = null;
-
-        constructor(_url: string) {
-          setTimeout(() => this.onopen?.(), 0);
-        }
-
-        send(data: string): void {
-          const msg = JSON.parse(data) as { type: string };
-          if (msg.type === "livekit_join") {
-            setTimeout(
-              () =>
-                this.onmessage?.({
-                  data: JSON.stringify({
-                    type: "livekit_room_info",
-                    livekit_url: "wss://livekit.example.com",
-                    token: "token",
-                    room_name: "room",
-                  }),
-                }),
-              0,
-            );
-          }
-        }
-
-        close(): void {
-          this.readyState = FakeWebSocket.CLOSED;
-          this.onclose?.({ code: 1000, reason: "" });
-        }
-      }
-
-      vi.stubGlobal("WebSocket", FakeWebSocket as unknown as typeof WebSocket);
-
-      try {
-        const onError = vi.fn();
-        const connection = new LiveKitConnection({ initialImage: "base64-image", onError });
-        const internal = connection as unknown as {
-          handleControlMessage: (msg: { type: "error"; error: string }) => void;
-          joinRoom: (info: unknown) => Promise<void>;
-        };
-        const joinSpy = vi.spyOn(internal, "joinRoom").mockResolvedValue(undefined);
-        const imageSpy = vi.spyOn(connection, "setImageBase64").mockReturnValue(new Promise<void>(() => {}));
-
-        const connectPromise = connection.connect("wss://example.com/v1/stream", {} as MediaStream, 750);
-        await vi.waitFor(() => expect(imageSpy).toHaveBeenCalled());
-
-        internal.handleControlMessage({ type: "error", error: "server unavailable" });
-
-        await expect(connectPromise).rejects.toThrow("server unavailable");
-        expect(onError).toHaveBeenCalledTimes(1);
-
-        joinSpy.mockRestore();
-        imageSpy.mockRestore();
-      } finally {
-        vi.unstubAllGlobals();
-      }
-    });
-  });
 });
 
 describe("Queue API", () => {
@@ -1320,104 +1160,37 @@ describe("Lucy 2.1 realtime", () => {
   });
 });
 
-describe("LiveKitConnection", () => {
-  describe("setImageBase64", () => {
-    it("rejects immediately when WebSocket is not open", async () => {
-      const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-      const connection = new LiveKitConnection();
+describe("Realtime Image Message Types", () => {
+  it("SetImageMessage has correct structure", () => {
+    const message: import("../src/realtime/types").SetImageMessage = {
+      type: "set_image",
+      image_data: "base64encodeddata",
+    };
 
-      await expect(connection.setImageBase64("base64data", { timeout: 5000 })).rejects.toThrow("WebSocket is not open");
-    });
+    expect(message.type).toBe("set_image");
+    expect(message.image_data).toBe("base64encodeddata");
+  });
 
-    it("rejects immediately with default timeout when WebSocket is not open", async () => {
-      const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-      const connection = new LiveKitConnection();
+  it("SetImageAckMessage has correct structure", () => {
+    const successMessage: import("../src/realtime/types").SetImageAckMessage = {
+      type: "set_image_ack",
+      success: true,
+      error: null,
+    };
 
-      await expect(connection.setImageBase64("base64data")).rejects.toThrow("WebSocket is not open");
-    });
+    expect(successMessage.type).toBe("set_image_ack");
+    expect(successMessage.success).toBe(true);
+    expect(successMessage.error).toBeNull();
 
-    describe("timeout behavior", () => {
-      beforeEach(() => {
-        vi.useFakeTimers();
-      });
+    const failureMessage: import("../src/realtime/types").SetImageAckMessage = {
+      type: "set_image_ack",
+      success: false,
+      error: "invalid image",
+    };
 
-      afterEach(() => {
-        vi.useRealTimers();
-      });
-
-      it("uses custom timeout when send succeeds but ack is not received", async () => {
-        const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-        const connection = new LiveKitConnection();
-        const sendSpy = vi.spyOn(connection, "send").mockReturnValue(true);
-
-        const customTimeout = 5000;
-        let rejected = false;
-        let rejectionError: Error | null = null;
-
-        const promise = connection.setImageBase64("base64data", { timeout: customTimeout }).catch((err) => {
-          rejected = true;
-          rejectionError = err;
-        });
-
-        await vi.advanceTimersByTimeAsync(customTimeout - 1);
-        expect(rejected).toBe(false);
-
-        await vi.advanceTimersByTimeAsync(2);
-        await promise;
-
-        expect(rejected).toBe(true);
-        expect(rejectionError?.message).toBe("Image send timed out");
-        sendSpy.mockRestore();
-      });
-
-      it("uses default timeout (30000ms) when send succeeds but ack is not received", async () => {
-        const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-        const connection = new LiveKitConnection();
-        const sendSpy = vi.spyOn(connection, "send").mockReturnValue(true);
-
-        let rejected = false;
-        let rejectionError: Error | null = null;
-
-        const promise = connection.setImageBase64("base64data").catch((err) => {
-          rejected = true;
-          rejectionError = err;
-        });
-
-        await vi.advanceTimersByTimeAsync(29999);
-        expect(rejected).toBe(false);
-
-        await vi.advanceTimersByTimeAsync(2);
-        await promise;
-
-        expect(rejected).toBe(true);
-        expect(rejectionError?.message).toBe("Image send timed out");
-        sendSpy.mockRestore();
-      });
-    });
-
-    it("sends set_image with null image_data and null prompt for passthrough", async () => {
-      vi.useFakeTimers();
-      try {
-        const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-        const connection = new LiveKitConnection();
-        const sendSpy = vi.spyOn(connection, "send").mockReturnValue(true);
-
-        const promise = connection.setImageBase64(null, { prompt: null }).catch(() => {});
-
-        expect(sendSpy).toHaveBeenCalledWith({
-          type: "set_image",
-          image_data: null,
-          prompt: null,
-        });
-
-        // Drain the ack timeout so no timers leak
-        await vi.advanceTimersByTimeAsync(30001);
-        await promise;
-        sendSpy.mockRestore();
-      } finally {
-        vi.useRealTimers();
-      }
-    });
+    expect(failureMessage.type).toBe("set_image_ack");
+    expect(failureMessage.success).toBe(false);
+    expect(failureMessage.error).toBe("invalid image");
   });
 });
 
@@ -1596,392 +1369,6 @@ describe("Subscribe Token", () => {
     const { decodeSubscribeToken } = await import("../src/realtime/subscribe-client.js");
     const token = btoa(JSON.stringify({ sid: "s" }));
     expect(() => decodeSubscribeToken(token)).toThrow("Invalid subscribe token");
-  });
-});
-
-describe("Subscribe Client", () => {
-  it("subscribe mode allows reconnect with null localStream", async () => {
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-
-    const manager = new LiveKitManager({
-      url: "wss://example.com",
-      onRemoteStream: vi.fn(),
-      onError: vi.fn(),
-    });
-
-    const internal = manager as unknown as {
-      handleConnectionStateChange: (state: import("../src/realtime/types").ConnectionState) => void;
-      reconnect: () => Promise<void>;
-      subscribeMode: boolean;
-      hasConnected: boolean;
-    };
-
-    internal.subscribeMode = true;
-    internal.hasConnected = true;
-
-    const reconnectSpy = vi.spyOn(internal, "reconnect").mockResolvedValue(undefined);
-    try {
-      internal.handleConnectionStateChange("disconnected");
-      expect(reconnectSpy).toHaveBeenCalledTimes(1);
-    } finally {
-      reconnectSpy.mockRestore();
-    }
-  });
-
-  it("publishes video with fixed h264 codec and 2.5Mbps bitrate", async () => {
-    const { LiveKitConnection } = await import("../src/realtime/livekit-connection.js");
-    const publishTrack = vi.fn().mockResolvedValue({ trackSid: "video-sid", mimeType: "video/H264" });
-    const videoTrack = {
-      kind: "video",
-      label: "camera",
-      getSettings: () => ({ width: 1280, height: 720, frameRate: 22 }),
-    } as unknown as MediaStreamTrack;
-    const stream = { getTracks: () => [videoTrack] } as unknown as MediaStream;
-    const connection = new LiveKitConnection();
-    const internal = connection as unknown as {
-      room: unknown;
-      publishLocalTracks: (stream: MediaStream) => Promise<void>;
-    };
-    internal.room = { localParticipant: { publishTrack } };
-
-    await internal.publishLocalTracks(stream);
-
-    expect(publishTrack).toHaveBeenCalledTimes(1);
-    const options = publishTrack.mock.calls[0]?.[1] as Record<string, unknown>;
-    expect(options).toMatchObject({
-      videoCodec: "h264",
-      videoEncoding: { maxBitrate: 2_500_000 },
-    });
-    expect(options).not.toHaveProperty("simulcast");
-    expect(options).not.toHaveProperty("scalabilityMode");
-    expect(options).not.toHaveProperty("degradationPreference");
-    expect(options.videoEncoding).not.toHaveProperty("maxFramerate");
-  });
-
-  it("session_id message populates subscribeToken on producer client", async () => {
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-    const { decodeSubscribeToken } = await import("../src/realtime/subscribe-client.js");
-
-    const sessionIdListeners = new Set<(msg: import("../src/realtime/types").SessionIdMessage) => void>();
-    const websocketEmitter = {
-      on: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.add(listener);
-      },
-      off: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.delete(listener);
-      },
-    };
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const mgr = this as unknown as {
-        config: { onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-      mgr.managerState = "connected";
-      mgr.config.onConnectionStateChange?.("connected");
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockReturnValue("connected");
-    const emitterSpy = vi
-      .spyOn(LiveKitManager.prototype, "getWebsocketMessageEmitter")
-      .mockReturnValue(websocketEmitter as never);
-    const sendSpy = vi.spyOn(LiveKitManager.prototype, "sendMessage").mockReturnValue(true);
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const realtime = createRealTimeClient({ baseUrl: "wss://api3.decart.ai", apiKey: "test-key" });
-      const client = await realtime.connect({} as MediaStream, {
-        model: models.realtime("mirage_v2"),
-        onRemoteStream: vi.fn(),
-      });
-
-      expect(client.subscribeToken).toBeNull();
-
-      for (const listener of sessionIdListeners) {
-        listener({
-          type: "session_id",
-          session_id: "sess-abc",
-          server_ip: "10.0.0.5",
-          server_port: 9090,
-        });
-      }
-
-      const token = client.subscribeToken;
-      expect(token).not.toBeNull();
-      // biome-ignore lint/style/noNonNullAssertion: guarded by assertion above
-      const decoded = decodeSubscribeToken(token!);
-      expect(decoded.sid).toBe("sess-abc");
-      expect(decoded.ip).toBe("10.0.0.5");
-      expect(decoded.port).toBe(9090);
-    } finally {
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      emitterSpy.mockRestore();
-      sendSpy.mockRestore();
-      cleanupSpy.mockRestore();
-    }
-  });
-
-  it("buffers pre-session telemetry diagnostics and discards them on disconnect", async () => {
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-
-    const fetchMock = vi.fn().mockResolvedValue({ ok: true });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const sessionIdListeners = new Set<(msg: import("../src/realtime/types").SessionIdMessage) => void>();
-    const websocketEmitter = {
-      on: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.add(listener);
-      },
-      off: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.delete(listener);
-      },
-    };
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const mgr = this as unknown as {
-        config: {
-          onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void;
-          onDiagnostic?: (name: string, data: unknown) => void;
-        };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-
-      mgr.config.onDiagnostic?.("phaseTiming", {
-        phase: "websocket",
-        durationMs: 12,
-        success: true,
-      });
-
-      mgr.managerState = "connected";
-      mgr.config.onConnectionStateChange?.("connected");
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockReturnValue("connected");
-    const emitterSpy = vi
-      .spyOn(LiveKitManager.prototype, "getWebsocketMessageEmitter")
-      .mockReturnValue(websocketEmitter as never);
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const realtime = createRealTimeClient({
-        baseUrl: "wss://api3.decart.ai",
-        apiKey: "test-key",
-        logger: { debug() {}, info() {}, warn() {}, error() {} },
-        telemetryEnabled: true,
-      });
-      const client = await realtime.connect({} as MediaStream, {
-        model: models.realtime("mirage_v2"),
-        onRemoteStream: vi.fn(),
-      });
-
-      expect(fetchMock).not.toHaveBeenCalled();
-
-      for (const listener of sessionIdListeners) {
-        listener({
-          type: "session_id",
-          session_id: "sess-telemetry",
-          server_ip: "10.0.0.5",
-          server_port: 9090,
-        });
-      }
-
-      // disconnect() calls stop() which discards buffered data instead of sending it.
-      // This avoids 401s when the API key has been invalidated on reconnect.
-      client.disconnect();
-
-      expect(fetchMock).not.toHaveBeenCalled();
-    } finally {
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      emitterSpy.mockRestore();
-      cleanupSpy.mockRestore();
-      vi.unstubAllGlobals();
-    }
-  });
-
-  it("stops previous telemetry reporter when session_id changes", async () => {
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-
-    vi.useFakeTimers();
-    const fetchMock = vi.fn().mockResolvedValue({ ok: true });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const setIntervalSpy = vi.spyOn(globalThis, "setInterval");
-    const clearIntervalSpy = vi.spyOn(globalThis, "clearInterval");
-
-    const sessionIdListeners = new Set<(msg: import("../src/realtime/types").SessionIdMessage) => void>();
-    const websocketEmitter = {
-      on: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.add(listener);
-      },
-      off: (event: string, listener: (msg: import("../src/realtime/types").SessionIdMessage) => void) => {
-        if (event === "sessionId") sessionIdListeners.delete(listener);
-      },
-    };
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const mgr = this as unknown as {
-        config: { onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-      mgr.managerState = "connected";
-      mgr.config.onConnectionStateChange?.("connected");
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockReturnValue("connected");
-    const emitterSpy = vi
-      .spyOn(LiveKitManager.prototype, "getWebsocketMessageEmitter")
-      .mockReturnValue(websocketEmitter as never);
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const realtime = createRealTimeClient({
-        baseUrl: "wss://api3.decart.ai",
-        apiKey: "test-key",
-        logger: { debug() {}, info() {}, warn() {}, error() {} },
-        telemetryEnabled: true,
-      });
-      const client = await realtime.connect({} as MediaStream, {
-        model: models.realtime("mirage_v2"),
-        onRemoteStream: vi.fn(),
-      });
-
-      for (const listener of sessionIdListeners) {
-        listener({
-          type: "session_id",
-          session_id: "sess-1",
-          server_ip: "10.0.0.5",
-          server_port: 9090,
-        });
-      }
-      for (const listener of sessionIdListeners) {
-        listener({
-          type: "session_id",
-          session_id: "sess-2",
-          server_ip: "10.0.0.6",
-          server_port: 9091,
-        });
-      }
-
-      client.disconnect();
-
-      expect(setIntervalSpy).toHaveBeenCalledTimes(2);
-      expect(clearIntervalSpy).toHaveBeenCalledTimes(2);
-    } finally {
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      emitterSpy.mockRestore();
-      cleanupSpy.mockRestore();
-      setIntervalSpy.mockRestore();
-      clearIntervalSpy.mockRestore();
-      vi.useRealTimers();
-      vi.unstubAllGlobals();
-    }
-  });
-
-  it("restarts stats collection when stats source changes after reconnect (LiveKit)", async () => {
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-    const { WebRTCStatsCollector } = await import("../src/realtime/webrtc-stats.js");
-
-    const firstStatsSource = { getStats: vi.fn() };
-    const secondStatsSource = { getStats: vi.fn() };
-    let currentStatsSource: typeof firstStatsSource = firstStatsSource;
-    let onConnectionStateChange: ((state: import("../src/realtime/types").ConnectionState) => void) | undefined;
-
-    const startSpy = vi.spyOn(WebRTCStatsCollector.prototype, "start").mockImplementation(() => {});
-    const stopSpy = vi.spyOn(WebRTCStatsCollector.prototype, "stop").mockImplementation(() => {});
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const mgr = this as unknown as {
-        config: { onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-      onConnectionStateChange = mgr.config.onConnectionStateChange;
-      mgr.managerState = "connected";
-      mgr.config.onConnectionStateChange?.("connected");
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockReturnValue("connected");
-    const statsProviderSpy = vi
-      .spyOn(LiveKitManager.prototype, "getStatsProvider")
-      .mockImplementation(() => currentStatsSource);
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const realtime = createRealTimeClient({
-        baseUrl: "wss://api3.decart.ai",
-        apiKey: "test-key",
-        logger: { debug() {}, info() {}, warn() {}, error() {} },
-        telemetryEnabled: true,
-      });
-      const client = await realtime.connect({} as MediaStream, {
-        model: models.realtime("mirage_v2"),
-        onRemoteStream: vi.fn(),
-      });
-
-      expect(startSpy).toHaveBeenCalledTimes(1);
-      expect(startSpy.mock.calls[0][0]).toBe(firstStatsSource);
-
-      currentStatsSource = secondStatsSource;
-      onConnectionStateChange?.("connected");
-
-      expect(startSpy).toHaveBeenCalledTimes(2);
-      expect(startSpy.mock.calls[1][0]).toBe(secondStatsSource);
-
-      client.disconnect();
-      expect(stopSpy).toHaveBeenCalled();
-    } finally {
-      startSpy.mockRestore();
-      stopSpy.mockRestore();
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      statsProviderSpy.mockRestore();
-      cleanupSpy.mockRestore();
-    }
-  });
-
-  it("subscribe client buffers events until returned", async () => {
-    const { encodeSubscribeToken } = await import("../src/realtime/subscribe-client.js");
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const mgr = this as unknown as {
-        config: { onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-      mgr.managerState = "connected";
-      mgr.config.onConnectionStateChange?.("connected");
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockReturnValue("connected");
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const token = encodeSubscribeToken("sess-123", "10.0.0.1", 8080);
-      const realtime = createRealTimeClient({ baseUrl: "wss://api3.decart.ai", apiKey: "sub-key" });
-      const client = await realtime.subscribe({
-        token,
-        onRemoteStream: vi.fn(),
-      });
-
-      const states: import("../src/realtime/types").ConnectionState[] = [];
-      client.on("connectionChange", (state) => states.push(state));
-
-      await new Promise((resolve) => setTimeout(resolve, 20));
-      expect(states).toEqual(["connected"]);
-
-      client.disconnect();
-    } finally {
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      cleanupSpy.mockRestore();
-    }
   });
 });
 
@@ -2718,121 +2105,6 @@ describe("TelemetryReporter", () => {
       expect(body.tags.model).toBeUndefined();
     } finally {
       vi.unstubAllGlobals();
-    }
-  });
-});
-
-describe("WebSockets Connection", () => {
-  it("treats generating as an established connection for reconnect decisions", async () => {
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-    const manager = new LiveKitManager({
-      url: "wss://example.com",
-      onRemoteStream: vi.fn(),
-      onError: vi.fn(),
-    });
-
-    const internal = manager as unknown as {
-      handleConnectionStateChange: (state: import("../src/realtime/types").ConnectionState) => void;
-      reconnect: () => Promise<void>;
-    };
-
-    const reconnectSpy = vi.spyOn(internal, "reconnect").mockResolvedValue(undefined);
-    try {
-      internal.handleConnectionStateChange("generating");
-      internal.handleConnectionStateChange("disconnected");
-      expect(reconnectSpy).toHaveBeenCalledTimes(1);
-    } finally {
-      reconnectSpy.mockRestore();
-    }
-  });
-
-  it("replays connection events emitted during connect before returning client", async () => {
-    const { createRealTimeClient } = await import("../src/realtime/client.js");
-    const { LiveKitManager } = await import("../src/realtime/livekit-manager.js");
-
-    const promptAckListeners = new Set<(msg: import("../src/realtime/types").PromptAckMessage) => void>();
-    const websocketEmitter = {
-      on: (event: string, listener: (msg: import("../src/realtime/types").PromptAckMessage) => void) => {
-        if (event === "promptAck") promptAckListeners.add(listener);
-      },
-      off: (event: string, listener: (msg: import("../src/realtime/types").PromptAckMessage) => void) => {
-        if (event === "promptAck") promptAckListeners.delete(listener);
-      },
-    };
-
-    const connectSpy = vi.spyOn(LiveKitManager.prototype, "connect").mockImplementation(async function () {
-      const manager = this as unknown as {
-        config: {
-          onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void;
-          initialPrompt?: { text: string; enhance?: boolean };
-        };
-        managerState: import("../src/realtime/types").ConnectionState;
-      };
-      manager.managerState = "connected";
-      manager.config.onConnectionStateChange?.("connected");
-
-      // Simulate initial prompt sent via WebSocket during connection setup
-      if (manager.config.initialPrompt) {
-        await new Promise((resolve) => setTimeout(resolve, 0));
-        manager.managerState = "generating";
-        manager.config.onConnectionStateChange?.("generating");
-      }
-
-      return true;
-    });
-    const stateSpy = vi.spyOn(LiveKitManager.prototype, "getConnectionState").mockImplementation(function () {
-      const manager = this as unknown as { managerState: import("../src/realtime/types").ConnectionState };
-      return manager.managerState ?? "connected";
-    });
-    const emitterSpy = vi
-      .spyOn(LiveKitManager.prototype, "getWebsocketMessageEmitter")
-      .mockReturnValue(websocketEmitter as never);
-    const sendSpy = vi.spyOn(LiveKitManager.prototype, "sendMessage").mockImplementation(function (message) {
-      if (message.type === "prompt") {
-        setTimeout(() => {
-          const manager = this as unknown as {
-            config: { onConnectionStateChange?: (state: import("../src/realtime/types").ConnectionState) => void };
-            managerState: import("../src/realtime/types").ConnectionState;
-          };
-          manager.managerState = "generating";
-          manager.config.onConnectionStateChange?.("generating");
-          for (const listener of promptAckListeners) {
-            listener({
-              type: "prompt_ack",
-              prompt: message.prompt,
-              success: true,
-              error: null,
-            });
-          }
-        }, 0);
-      }
-      return true;
-    });
-    const cleanupSpy = vi.spyOn(LiveKitManager.prototype, "cleanup").mockImplementation(() => {});
-
-    try {
-      const realtime = createRealTimeClient({ baseUrl: "wss://example.com", apiKey: "test-key" });
-      const client = await realtime.connect({} as MediaStream, {
-        model: models.realtime("mirage_v2"),
-        onRemoteStream: vi.fn(),
-        initialState: {
-          prompt: {
-            text: "test",
-          },
-        },
-      });
-
-      const states: import("../src/realtime/types").ConnectionState[] = [];
-      client.on("connectionChange", (state) => states.push(state));
-
-      await new Promise((resolve) => setTimeout(resolve, 20));
-      expect(states).toEqual(["connected", "generating"]);
-    } finally {
-      connectSpy.mockRestore();
-      stateSpy.mockRestore();
-      emitterSpy.mockRestore();
-      sendSpy.mockRestore();
-      cleanupSpy.mockRestore();
     }
   });
 });
